requirements.txt
praw==7.7.1
langdetect==1.0.9
mysql-connector-python==8.4.0